Makes active_jobs JS code one file
