+++ apps/dashboard/app/views/active_jobs/index.html.erb
@@ -1,5 +1,3 @@
-<%=  javascript_include_tag 'active_jobs/active_jobs', nonce: true %>
-
 <div class="row">
   <div class="col-md-12">
     <div class="float-right">
@@ -68,4 +66,4 @@
   data-git-string=<%= 'git describe --always --tags'.strip %>
 ></div>
 
-<%=  javascript_include_tag 'active_jobs/performance_report', nonce: true %>
+<%=  javascript_include_tag 'active_jobs', nonce: true %>
